parseInt(matchAmount) to get rid of object Object error

+++ app/src/components/Forms/MatchContribution/MatchContribution.js
@@ -80,7 +80,7 @@ const MatchContribution = ({
               updateContribution({
                 id,
                 status: ContributionStatusEnum.PROCESSED,
-                matchAmount: parseFloat(matchAmount),
+                matchAmount: parseInt(matchAmount),
                 compliant: true,
               });
               postContributionComment(id, matchAmount);
